refactor(ios): organize mobile app structure
