+++ src/apps/mobile/ios/BitFun/App/BitFunApp.swift
@@ -2,9 +2,9 @@ import SwiftUI
 
 @main
 struct BitFunApp: App {
-    @StateObject private var model = MobileAppModel.launchConfigured
+    @StateObject private var model = MobileLaunchConfiguration.makeModel()
     @Environment(\.scenePhase) private var scenePhase
-    private let designPreviewScenario = Self.resolveDesignPreviewScenario()
+    private let designPreviewScenario = MobileLaunchConfiguration.designPreviewScenario()
 
     var body: some Scene {
         WindowGroup {
@@ -19,17 +19,4 @@ struct BitFunApp: App {
         }
     }
 
-    private static func resolveDesignPreviewScenario() -> MobilePreviewScenario? {
-        let arguments = ProcessInfo.processInfo.arguments
-        guard let marker = arguments.firstIndex(of: "--design-preview") else { return nil }
-        let scenarioID = arguments.indices.contains(marker + 1) ? arguments[marker + 1] : "connected-conversation"
-        switch scenarioID {
-        case MobilePreviewScenarios.streamingDark.id:
-            return MobilePreviewScenarios.streamingDark
-        case MobilePreviewScenarios.reconnectingWide.id:
-            return MobilePreviewScenarios.reconnectingWide
-        default:
-            return MobilePreviewScenarios.connectedConversation
-        }
-    }
 }
