+++ src/apps/mobile/ios/BitFun/Features/Chat/ConversationHomeViews.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+struct LocalHomeView: View {
+    var body: some View {
+        BitFunTheme.page
+    }
+}
