+++ src/apps/mobile/ios/BitFun/Features/Shell/MobileShellView.swift
@@ -1,4 +1,3 @@
-import AVFoundation
 import BitFunMobileCore
 import SwiftUI
 import UniformTypeIdentifiers
@@ -377,408 +376,5 @@ struct MobileShellView: View {
         if width > 0 {
             Rectangle().fill(BitFunTheme.line).frame(width: width)
         }
-    }
-}
-
-
-
-
-private struct LocalHomeView: View {
-    var body: some View {
-        BitFunTheme.page
-    }
-}
-
-private struct RemoteHomeView: View {
-    @ObservedObject var model: MobileAppModel
-
-    var body: some View {
-        ZStack(alignment: .topTrailing) {
-            VStack(spacing: 12) {
-            Spacer()
-            ZStack {
-                Image(systemName: "desktopcomputer")
-                    .font(.system(size: 42, weight: .medium))
-                    .foregroundStyle(BitFunTheme.ink)
-            }
-            .frame(width: 74, height: 74)
-            .background(BitFunTheme.card)
-            .overlay(RoundedRectangle(cornerRadius: 24).stroke(BitFunTheme.line, lineWidth: 1))
-            .clipShape(RoundedRectangle(cornerRadius: 24))
-            Text(model.localized("连接桌面端"))
-                .font(.system(size: 18, weight: .bold))
-                .foregroundStyle(BitFunTheme.ink)
-            Text(model.localized("扫描桌面端显示的二维码，开始远程处理任务。"))
-                .font(.system(size: 13))
-                .foregroundStyle(BitFunTheme.muted)
-                .multilineTextAlignment(.center)
-                .lineSpacing(7)
-                .padding(.horizontal, 20)
-            Button(model.localized("连接")) { model.connectRemote() }
-                .font(.system(size: 15, weight: .medium))
-                .foregroundStyle(.white)
-                .frame(width: 136, height: 44)
-                .background(BitFunTheme.accent)
-                .clipShape(Capsule())
-            Spacer()
-            }
-            remoteSettingsButton
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .padding(.horizontal, 20)
-        .padding(.bottom, 48)
-        .background(BitFunTheme.page)
-    }
-
-    private var remoteSettingsButton: some View {
-        Button { model.remoteControlSettingsOpen = true } label: {
-            Image(systemName: "gearshape")
-                .font(.system(size: 18, weight: .medium))
-                .foregroundStyle(BitFunTheme.ink)
-                .frame(width: 44, height: 44)
-                .background(BitFunTheme.card)
-                .overlay(Circle().stroke(BitFunTheme.line, lineWidth: 1))
-                .clipShape(Circle())
-        }
-        .buttonStyle(.plain)
-        .accessibilityLabel(model.localized("远程控制设置"))
-        .padding(.top, 16).padding(.trailing, 16)
-    }
-}
-
-private struct RemoteConnectedHomeView: View {
-    @ObservedObject var model: MobileAppModel
-
-    var body: some View {
-        ZStack(alignment: .topTrailing) {
-            VStack(spacing: 14) {
-            Spacer()
-            Image(systemName: "desktopcomputer.and.macbook")
-                .font(.system(size: 34, weight: .medium)).foregroundStyle(BitFunTheme.muted)
-            Text(model.localized("桌面端已连接"))
-                .font(MobileDesignTypography.titleMedium.font).foregroundStyle(BitFunTheme.ink)
-            Text(model.localized("选择已有会话，或在当前工作区创建一个新会话。"))
-                .font(MobileDesignTypography.bodySmall.font).foregroundStyle(BitFunTheme.muted)
-                .multilineTextAlignment(.center)
-            Button { model.remoteCreateOpen = true } label: {
-                Label(model.localized("新建远程会话"), systemImage: "plus")
-                    .font(MobileDesignTypography.labelMedium.font).foregroundStyle(.white)
-                    .frame(minWidth: 176, minHeight: 44).background(BitFunTheme.accent).clipShape(Capsule())
-            }
-            .buttonStyle(.plain)
-            Spacer()
-            }
-            Button { model.remoteControlSettingsOpen = true } label: {
-                Image(systemName: "gearshape")
-                    .font(.system(size: 18, weight: .medium))
-                    .foregroundStyle(BitFunTheme.ink)
-                    .frame(width: 44, height: 44)
-                    .background(BitFunTheme.card)
-                    .overlay(Circle().stroke(BitFunTheme.line, lineWidth: 1))
-                    .clipShape(Circle())
-            }
-            .buttonStyle(.plain)
-            .accessibilityLabel(model.localized("远程控制设置"))
-            .padding(.top, 16).padding(.trailing, 16)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .background(BitFunTheme.page)
-    }
-}
-
-private struct ConnectionStatusBar: View {
-    let phase: ConnectionPhase
-    var detail: String?
-    let onRetry: () -> Void
-    var body: some View {
-        HStack(spacing: 8) {
-            Circle().fill(phase == .reconnecting ? BitFunTheme.muted : BitFunTheme.red).frame(width: 8, height: 8)
-            Text(MobileLocalization.text(phase == .reconnecting ? "正在恢复连接" : "连接不可用"))
-                .font(.system(size: 13, weight: .medium))
-            Text(
-                detail ?? MobileLocalization.text(
-                    phase == .reconnecting ? "正在重新连接桌面端" : "请重新连接"
-                )
-            )
-                .font(.system(size: 12))
-                .foregroundStyle(BitFunTheme.muted)
-            Spacer()
-            if phase == .disconnected {
-                Button(MobileLocalization.text("重试"), action: onRetry)
-                    .font(.system(size: 13, weight: .semibold))
-                    .buttonStyle(.plain)
-                    .foregroundStyle(BitFunTheme.accent)
-            }
-        }
-        .foregroundStyle(BitFunTheme.ink)
-        .padding(.horizontal, 18)
-        .frame(height: 48)
-        .background(BitFunTheme.soft)
-    }
-}
-
-private struct SettingsView: View {
-    @ObservedObject var model: MobileAppModel
-    @Environment(\.dismiss) private var dismiss
-    @State private var accountOpen = false
-
-    private var appVersion: String {
-        Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String ?? "1.0.0"
-    }
-
-    private var selectedModelName: String {
-        model.modelOptions.first(where: \.selected)?.primaryLabel
-            ?? model.modelOptions.first?.primaryLabel
-            ?? model.localized("未配置")
-    }
-
-    var body: some View {
-        ZStack(alignment: .topTrailing) {
-            ScrollView(showsIndicators: false) {
-                VStack(alignment: .leading, spacing: 0) {
-                    Text(model.localized("设置"))
-                        .font(.system(size: 28, weight: .bold))
-                        .foregroundStyle(BitFunTheme.ink)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .padding(.bottom, 30)
-
-                    Button { accountOpen = true } label: {
-                        SettingsCard {
-                            SettingsProfileRow(
-                                subtitle: model.accountUser ?? model.localized("未登录")
-                            )
-                        }
-                    }
-                    .buttonStyle(.plain)
-                    .padding(.bottom, 24)
-
-                    SettingsGroup(title: "通用") {
-                        VStack(spacing: 0) {
-                            Button { model.languagePickerOpen = true } label: {
-                                SettingsValueRow(
-                                    icon: "textformat",
-                                    title: "语言",
-                                    value: model.appLanguage.nativeName,
-                                    showsChevron: true
-                                )
-                            }
-                            .buttonStyle(.plain)
-                            Divider().overlay(BitFunTheme.line).padding(.horizontal, 26)
-                            Button { model.generalConfigOpen = true } label: {
-                                SettingsValueRow(
-                                    icon: "square.grid.2x2",
-                                    title: "模型",
-                                    value: selectedModelName,
-                                    showsChevron: true
-                                )
-                            }
-                            .buttonStyle(.plain)
-                        }
-                    }
-                    SettingsGroup(title: "关于") {
-                        VStack(spacing: 0) {
-                            SettingsValueRow(
-                                icon: nil,
-                                title: "产品",
-                                value: "BitFun iOS版"
-                            )
-                            Divider().overlay(BitFunTheme.line).padding(.horizontal, 26)
-                            SettingsValueRow(icon: nil, title: "版本", value: appVersion)
-                        }
-                    }
-                }
-                .padding(.horizontal, 16)
-                .padding(.top, 64)
-                .padding(.bottom, 34)
-            }
-
-            Button { dismiss() } label: {
-                Image(systemName: "xmark")
-                    .font(.system(size: 18, weight: .regular))
-                    .foregroundStyle(BitFunTheme.ink)
-                    .frame(width: 40, height: 40)
-                    .background(BitFunTheme.card)
-                    .clipShape(Circle())
-            }
-            .buttonStyle(.plain)
-            .accessibilityLabel(model.localized("关闭"))
-            .padding(.top, 22)
-            .padding(.trailing, 18)
-
-            if model.languagePickerOpen {
-                LanguagePickerSheet(model: model)
-                    .transition(.move(edge: .trailing).combined(with: .opacity))
-            } else if model.generalConfigOpen {
-                GeneralChatConfigSheet(model: model)
-                    .transition(.move(edge: .trailing).combined(with: .opacity))
-            } else if accountOpen {
-                AccountSettingsView(model: model, onClose: { accountOpen = false })
-                    .transition(.move(edge: .trailing).combined(with: .opacity))
-            }
-        }
-        .background(BitFunTheme.page)
-        .animation(.easeInOut(duration: 0.2), value: model.languagePickerOpen)
-        .animation(.easeInOut(duration: 0.2), value: model.generalConfigOpen)
-        .animation(.easeInOut(duration: 0.2), value: accountOpen)
-    }
-}
-
-private struct LanguagePickerSheet: View {
-    @ObservedObject var model: MobileAppModel
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            BitFunSelectionHeader(title: "选择语言", onClose: { model.languagePickerOpen = false })
-            Divider().overlay(BitFunTheme.line)
-
-            VStack(spacing: 0) {
-                ForEach(MobileLanguage.allCases) { language in
-                    Button {
-                        model.setLanguage(language)
-                        model.languagePickerOpen = false
-                    } label: {
-                        HStack {
-                            Text(language.nativeName)
-                                .font(.system(size: 16, weight: .medium))
-                                .foregroundStyle(BitFunTheme.ink)
-                            Spacer()
-                            if model.appLanguage == language {
-                                Image(systemName: "checkmark")
-                                    .font(.system(size: 18, weight: .medium))
-                                    .foregroundStyle(BitFunTheme.ink)
-                            }
-                        }
-                        .padding(.horizontal, 16)
-                        .frame(height: MobileDesignGeometry.selectionRowHeight)
-                        .contentShape(Rectangle())
-                    }
-                    .buttonStyle(.plain)
-                }
-            }
-            .padding(.top, 8)
-            .padding(.bottom, 28)
-
-            Spacer(minLength: 0)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .top)
-        .background(BitFunTheme.card)
-        .clipShape(RoundedRectangle(cornerRadius: MobileDesignGeometry.selectionTopRadius))
-    }
-}
-
-
-
-
-
-private struct PermissionModeRow: View {
-    @ObservedObject var model: MobileAppModel
-    let mode: String
-    let title: String
-    let detail: String
-
-    var body: some View {
-        Button { model.setRemotePermissionMode(mode) } label: {
-            HStack(spacing: 12) {
-                VStack(alignment: .leading, spacing: 3) {
-                    Text(model.localized(title)).font(MobileDesignTypography.bodyMedium.font).foregroundStyle(BitFunTheme.ink)
-                    Text(model.localized(detail)).font(MobileDesignTypography.labelSmall.font).foregroundStyle(BitFunTheme.muted)
-                }
-                Spacer()
-                if model.remotePermissionMode == mode {
-                    Image(systemName: "checkmark.circle.fill").foregroundStyle(BitFunTheme.green)
-                }
-            }
-            .padding(.horizontal, 20).frame(minHeight: 62)
-        }
-        .buttonStyle(.plain).disabled(model.busy)
-    }
-}
-
-private struct SettingsGroup<Content: View>: View {
-    let title: String
-    @ViewBuilder let content: () -> Content
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text(MobileLocalization.text(title))
-                .font(.system(size: 18, weight: .bold))
-                .foregroundStyle(BitFunTheme.muted)
-                .padding(.leading, 12)
-            SettingsCard(content: content)
-        }
-        .padding(.bottom, 24)
-    }
-}
-
-struct SettingsCard<Content: View>: View {
-    @ViewBuilder let content: () -> Content
-
-    var body: some View {
-        BitFunModalCard(
-            radius: MobileDesignGeometry.settingsCompactCardRadius,
-            bordered: false,
-            content: content
-        )
-    }
-}
-
-private struct SettingsProfileRow: View {
-    let subtitle: String
-
-    var body: some View {
-        HStack(spacing: 12) {
-            Image(systemName: "person.crop.circle")
-                .font(.system(size: 24, weight: .regular))
-                .foregroundStyle(BitFunTheme.muted)
-                .frame(width: 34, height: 34)
-            VStack(alignment: .leading, spacing: 2) {
-                Text(MobileLocalization.text("个人资料"))
-                    .font(.system(size: 16, weight: .medium))
-                    .foregroundStyle(BitFunTheme.ink)
-                Text(MobileLocalization.text(subtitle))
-                    .font(.system(size: 13))
-                    .foregroundStyle(BitFunTheme.muted)
-                    .lineLimit(1)
-            }
-            Spacer(minLength: 8)
-            Image(systemName: "chevron.right")
-                .font(.system(size: 14, weight: .medium))
-                .foregroundStyle(BitFunTheme.muted.opacity(0.72))
-        }
-        .padding(.horizontal, 18)
-        .frame(height: 64)
-    }
-}
-
-private struct SettingsValueRow: View {
-    let icon: String?
-    let title: String
-    let value: String
-    var showsChevron: Bool = false
-
-    var body: some View {
-        HStack(spacing: 14) {
-            if let icon {
-                Image(systemName: icon)
-                    .font(.system(size: 20, weight: .regular))
-                    .foregroundStyle(BitFunTheme.muted)
-                    .frame(width: 23, height: 23)
-            }
-            Text(MobileLocalization.text(title))
-                .font(.system(size: 16, weight: .medium))
-                .foregroundStyle(BitFunTheme.ink)
-            Spacer(minLength: 12)
-            Text(MobileLocalization.text(value))
-                .font(.system(size: 15))
-                .foregroundStyle(BitFunTheme.muted)
-                .lineLimit(1)
-            if showsChevron {
-                Image(systemName: "chevron.right")
-                    .font(.system(size: 14, weight: .medium))
-                    .foregroundStyle(BitFunTheme.muted.opacity(0.72))
-            }
-        }
-        .padding(.horizontal, 18)
-        .frame(height: 52)
     }
 }
