+++ src/apps/mobile/ios/BitFun/Infrastructure/MobileAppModel.swift
@@ -2,291 +2,6 @@ import Foundation
 import SwiftUI
 import BitFunMobileCore
 
-enum ConnectionPhase {
-    case connected
-    case reconnecting
-    case disconnected
-}
-
-enum MobileSurface: String {
-    case local
-    case remote
-}
-
-struct ChatMessage: Identifiable, Equatable {
-    let id: UUID
-    let role: Role
-    let text: String
-
-    enum Role { case user, assistant }
-}
-
-struct MobileTimelineImage: Identifiable, Equatable {
-    var id: String { dataURL }
-    let name: String
-    let dataURL: String
-}
-
-struct MobileTimelineOption: Identifiable, Equatable {
-    let label: String
-    let description: String?
-    var id: String { label }
-}
-
-struct MobileTimelineQuestion: Identifiable, Equatable {
-    let index: Int
-    let header: String
-    let question: String
-    let options: [MobileTimelineOption]
-    let multiSelect: Bool
-    var id: Int { index }
-}
-
-struct MobileTimelineTool: Identifiable, Equatable {
-    let id: String
-    let name: String
-    let phase: String
-    let kind: String
-    let operation: String
-    let target: String
-    let filePath: String
-    let fileLabel: String
-    let input: String
-    let output: String
-    let question: String?
-    let questions: [MobileTimelineQuestion]
-    let actions: Set<String>
-}
-
-indirect enum MobileTimelineBlock: Identifiable, Equatable {
-    case text(id: String, text: String, streaming: Bool)
-    case thinking(id: String, text: String, streaming: Bool)
-    case tools(id: String, tools: [MobileTimelineTool])
-    case subagent(
-        id: String,
-        title: String,
-        running: Bool,
-        text: String,
-        children: [MobileTimelineBlock]
-    )
-
-    var id: String {
-        switch self {
-        case let .text(id, _, _), let .thinking(id, _, _), let .tools(id, _),
-             let .subagent(id, _, _, _, _):
-            return id
-        }
-    }
-}
-
-struct MobileConversationRow: Identifiable, Equatable {
-    let id: String
-    let kind: String
-    let text: String
-    let thinking: String?
-    let images: [MobileTimelineImage]
-    let tools: [MobileTimelineTool]
-    let blocks: [MobileTimelineBlock]
-    let streaming: Bool
-    let typing: Bool
-    let pending: Bool
-    let showRetry: Bool
-}
-
-enum MobileFilePreviewFailureKind: String {
-    case notFound, unavailable, accessDenied, tooLarge, connection, loadFailed
-}
-
-struct MobileFilePreview: Identifiable, Equatable {
-    let id: String
-    let sessionID: String
-    let controlTargetEpoch: Int32
-    let name: String
-    let content: String
-    let mimeType: String
-    let imageData: Data?
-    let truncated: Bool
-    let loadedBytes: Int64
-    let sizeBytes: Int64
-    let markdown: Bool
-    let lineStart: Int32
-    let failure: String?
-    let failureKind: MobileFilePreviewFailureKind?
-    let retryable: Bool
-    let unsupported: Bool
-
-    init(
-        id: String,
-        sessionID: String = "",
-        controlTargetEpoch: Int32 = 0,
-        name: String,
-        content: String,
-        mimeType: String,
-        imageData: Data?,
-        truncated: Bool,
-        loadedBytes: Int64 = 0,
-        sizeBytes: Int64 = 0,
-        markdown: Bool = false,
-        lineStart: Int32 = 0,
-        failure: String?,
-        failureKind: MobileFilePreviewFailureKind? = nil,
-        retryable: Bool = false,
-        unsupported: Bool = false
-    ) {
-        self.id = id
-        self.sessionID = sessionID
-        self.controlTargetEpoch = controlTargetEpoch
-        self.name = name
-        self.content = content
-        self.mimeType = mimeType
-        self.imageData = imageData
-        self.truncated = truncated
-        self.loadedBytes = loadedBytes
-        self.sizeBytes = sizeBytes
-        self.markdown = markdown
-        self.lineStart = lineStart
-        self.failure = failure
-        self.failureKind = failureKind
-        self.retryable = retryable
-        self.unsupported = unsupported
-    }
-}
-
-struct MobilePendingDownload: Identifiable, Equatable {
-    var id: String { reference }
-    let reference: String
-    let remotePath: String
-    let name: String
-    let mimeType: String
-    let data: Data
-    let sessionID: String
-    let controlTargetEpoch: Int32
-
-    init(reference: String, remotePath: String, name: String, mimeType: String, data: Data,
-         sessionID: String = "", controlTargetEpoch: Int32 = 0) {
-        self.reference = reference
-        self.remotePath = remotePath
-        self.name = name
-        self.mimeType = mimeType
-        self.data = data
-        self.sessionID = sessionID
-        self.controlTargetEpoch = controlTargetEpoch
-    }
-}
-
-struct ChatSession: Identifiable, Equatable {
-    let id: String
-    var title: String
-    var updatedLabel: String
-    var pinned: Bool = false
-    var status: String = "active"
-    var agentType: String = "general_chat"
-    var workspacePath: String?
-    var workspaceName: String?
-    var deviceKey: String? = nil
-    var createdAt: String = ""
-    var messageCount: Int = 0
-}
-
-struct CommittedRemoteCreate {
-    let targetKey: String
-    let epoch: UInt64
-    let session: ChatSession
-    /// First authoritative Ready revision guaranteed to contain this commit.
-    let minimumAuthorityRevision: Int64
-}
-
-struct PendingDirectoryRemoteDraft {
-    let targetKey: String
-    let rawDeviceKey: String
-    let workspacePath: String
-    let normalizedWorkspacePath: String
-    let epoch: UInt64
-    var selectionRequested: Bool
-}
-
-struct MobileAccountDevice: Identifiable, Equatable {
-    let id: String
-    let name: String
-    let online: Bool
-    let selected: Bool
-}
-
-struct MobileDeviceDirectoryEntry: Identifiable, Equatable {
-    let id: String
-    let name: String
-    let online: Bool
-    let expanded: Bool
-    let status: String
-    let error: String?
-    let workspaces: [MobileWorkspaceGroup]
-    let sessions: [ChatSession]
-}
-
-struct MobileWorkspaceGroup: Identifiable, Equatable {
-    var id: String { (deviceKey ?? "") + ":" + path }
-    let path: String
-    let name: String
-    let selected: Bool
-    let sessions: [ChatSession]
-    var deviceKey: String? = nil
-}
-
-enum MobileSessionListSectionKind: Equatable {
-    case chat
-    case project
-    case today
-    case yesterday
-    case earlier
-}
-
-struct MobileSessionListSectionProjection: Identifiable {
-    let id: String
-    let kind: MobileSessionListSectionKind
-    let path: String
-    let name: String
-    let sessions: [ChatSession]
-}
-
-struct MobileSessionWorkspaceOption: Identifiable {
-    var id: String { path }
-    let path: String
-    let name: String
-}
-
-struct MobileAssistantOption: Identifiable, Equatable {
-    var id: String { path }
-    let path: String
-    let name: String
-}
-
-struct ComposerAttachment: Identifiable, Equatable {
-    let id: String
-    let data: Data
-    let mimeType: String
-
-    var dataURL: String {
-        "data:\(mimeType);base64,\(data.base64EncodedString())"
-    }
-}
-
-struct ComposerModelOption: Identifiable, Equatable {
-    let id: String
-    let primaryLabel: String
-    let secondaryLabel: String
-    let source: String
-    let selected: Bool
-}
-
-enum MobileDownloadPhase {
-    case idle
-    case preparing
-    case downloading
-    case saving
-    case saved
-    case failed
-}
-
 @MainActor
 final class MobileAppModel: ObservableObject {
     @Published var appLanguage: MobileLanguage = MobileLocalization.restoredLanguage()
@@ -442,194 +157,6 @@ final class MobileAppModel: ObservableObject {
         self.localDeviceID = adapter.deviceID
     }
 
-    static let preview: MobileAppModel = {
-        let first = ChatSession(id: UUID().uuidString, title: "你好", updatedLabel: "刚刚")
-        return MobileAppModel(
-            sessions: [first],
-            selectedSessionID: first.id,
-            messages: [
-                ChatMessage(id: UUID(), role: .user, text: "你好"),
-                ChatMessage(id: UUID(), role: .assistant, text: "这是 BitFun 的移动端会话界面。你可以从手机连接桌面端，查看工作区、会话和 Agent 的执行状态。")
-            ]
-        )
-    }()
-
-    static var launchConfigured: MobileAppModel {
-        let model = preview
-        let arguments = ProcessInfo.processInfo.arguments
-        if arguments.contains("--english") {
-            model.setLanguage(.english)
-        } else if arguments.contains("--simplified-chinese") {
-            model.setLanguage(.simplifiedChinese)
-        }
-        if arguments.contains("--remote") {
-            model.surface = .remote
-        }
-        if arguments.contains("--connected") {
-            model.configureConnectedPreview()
-        }
-        if arguments.contains("--remote-chat-section") {
-            if !model.remoteConnected { model.configureConnectedPreview() }
-            model.remoteSessions.append(
-                ChatSession(
-                    id: "preview-remote-chat",
-                    title: "移动端体验对齐",
-                    updatedLabel: "刚刚",
-                    status: "idle",
-                    agentType: "Claw",
-                    workspacePath: nil,
-                    workspaceName: nil
-                )
-            )
-            model.rebuildRemoteWorkspaceGroups()
-        }
-        if arguments.contains("--remote-view-settings") {
-            if !model.remoteConnected { model.configureConnectedPreview() }
-            model.remoteViewSettingsOpen = true
-        }
-        if arguments.contains("--remote-view-density") {
-            if !model.remoteConnected { model.configureConnectedPreview() }
-            let now = ISO8601DateFormatter().string(from: Date())
-            for index in model.remoteSessions.indices {
-                model.remoteSessions[index].updatedLabel = now
-            }
-            model.remoteGroupMode = "TIME"
-            model.remoteShowWorkspaceMetadata = true
-            model.remoteShowUpdatedMetadata = true
-            model.remoteShowStatusMetadata = true
-            model.rebuildRemoteWorkspaceGroups()
-        }
-        if arguments.contains("--timeline-preview") {
-            model.configureTimelinePreview()
-        }
-        if arguments.contains("--file-preview") {
-            model.filePreview = MobileFilePreview(
-                id: "src/main.rs",
-                name: "main.rs",
-                content: "// Remote workspace preview\nfn main() {\n    println!(\"Hello from BitFun\");\n}\n",
-                mimeType: "text/x-rust",
-                imageData: nil,
-                truncated: false,
-                failure: nil
-            )
-        }
-        if arguments.contains("--download-preview") {
-            model.pendingDownload = MobilePendingDownload(
-                reference: "computer://src/main.rs",
-                remotePath: "src/main.rs",
-                name: "main.rs",
-                mimeType: "text/x-rust",
-                data: Data("fn main() {}\n".utf8)
-            )
-            model.downloadTargetPath = "src/main.rs"
-            model.downloadPhase = .saving
-            model.downloadStatusText = model.localized("正在保存")
-            model.downloadExporterOpen = true
-        }
-        if let relay = arguments.value(after: "--relay-url"),
-           let username = arguments.value(after: "--username"),
-           let password = arguments.value(after: "--password") {
-            model.loginAccount(relayURL: relay, username: username, password: password)
-        }
-        if arguments.contains("--drawer") {
-            model.drawerOpen = true
-        }
-        if arguments.contains("--settings") {
-            model.settingsOpen = true
-        }
-        if arguments.contains("--remote-settings") {
-            model.surface = .remote
-            model.remoteControlSettingsOpen = true
-        }
-        if arguments.contains("--model-settings") {
-            model.settingsOpen = true
-            model.generalConfigOpen = true
-        }
-        if arguments.contains("--composer-model-picker") ||
-            ProcessInfo.processInfo.environment["BITFUN_COMPOSER_MODEL_PICKER"] == "1" {
-            model.composerModelPickerPreview = true
-            model.localSessionSelected = true
-            model.draft = "\n"
-            model.modelOptions = [
-                ComposerModelOption(
-                    id: "preview-codex",
-                    primaryLabel: "GPT-5.6 Codex",
-                    secondaryLabel: "BitFun 账号",
-                    source: "ACCOUNT",
-                    selected: true
-                ),
-                ComposerModelOption(
-                    id: "preview-local",
-                    primaryLabel: "本机自定义模型",
-                    secondaryLabel: "OpenAI 兼容服务",
-                    source: "LOCAL",
-                    selected: false
-                ),
-            ]
-        }
-        if arguments.contains("--pairing") || arguments.contains("--pairing-manual") ||
-            arguments.contains("--pairing-account") {
-            model.pairingSheetOpen = true
-        }
-        if arguments.contains("--remote-create") || arguments.contains("--remote-create-workspace-picker") {
-            model.remoteCreatePreview = true
-            if !model.remoteConnected { model.configureConnectedPreview() }
-            model.remoteCreateOpen = true
-        }
-        if arguments.contains("--remote-home-preview") {
-            model.remoteSessionSelected = false
-            model.selectedSessionID = ""
-            model.timelineRows = []
-            model.messages = []
-        }
-        if arguments.contains("--local-actions") {
-            model.localActionPreview = true
-            model.surface = .local
-            model.localSessionSelected = true
-            model.remoteSessionSelected = false
-            if let localSession = model.sessions.first {
-                model.selectedSessionID = localSession.id
-            }
-        }
-        if arguments.contains("--account-login") {
-            model.accountLoginPreview = true
-            model.accountUser = nil
-            model.accountDeviceName = nil
-            model.accountSelectedDeviceID = nil
-            model.accountDevices = []
-            model.accountDeviceCount = 0
-            model.coreErrorMessage = nil
-            model.settingsOpen = false
-            model.accountSheetOpen = true
-        }
-        if arguments.contains("--account-profile") {
-            model.accountLoginPreview = true
-            model.accountUser = "bitfun-user"
-            model.accountUserID = "user-preview-7A31"
-            model.accountDevices = [
-                MobileAccountDevice(
-                    id: "desktop-preview",
-                    name: "Studio Mac",
-                    online: true,
-                    selected: true
-                ),
-                MobileAccountDevice(
-                    id: "desktop-offline-preview",
-                    name: "Office PC",
-                    online: false,
-                    selected: false
-                ),
-            ]
-            model.accountDeviceName = "Studio Mac"
-            model.accountSelectedDeviceID = "desktop-preview"
-            model.accountDeviceCount = model.accountDevices.count
-            model.coreErrorMessage = nil
-            model.settingsOpen = false
-            model.accountSheetOpen = true
-        }
-        return model
-    }
-
     var selectedSession: ChatSession? {
         guard (surface == .local && localSessionSelected) || (surface == .remote && remoteSessionSelected) else {
             return nil
@@ -640,8 +167,6 @@ final class MobileAppModel: ObservableObject {
     var visibleSessions: [ChatSession] {
         surface == .local ? sessions : remoteSessions
     }
-
-
 
     func switchSurface(_ next: MobileSurface) {
         surface = next
@@ -748,150 +273,12 @@ final class MobileAppModel: ObservableObject {
         drawerOpen = false
     }
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
     func showSessionDetails(_ session: ChatSession) {
         sessionDetails = session
     }
 
     func dismissSessionDetails() {
         sessionDetails = nil
-    }
-
-
-    private func configureConnectedPreview() {
-        directPairingConnected = true
-        surface = .remote
-        remoteConnected = true
-        connectionPhase = .connected
-        remoteSessionSelected = true
-        accountUser = "preview@bitfun"
-        accountDeviceName = "DESKTOP-KM3L4UI"
-        accountSelectedDeviceID = "preview-desktop"
-        directoryFixturePreview = true
-        accountDevices = [
-            MobileAccountDevice(id: "preview-desktop", name: "DESKTOP-KM3L4UI", online: true, selected: true),
-            MobileAccountDevice(id: "preview-mac", name: "Studio Mac", online: true, selected: false),
-            MobileAccountDevice(id: "preview-offline", name: "Office PC", online: false, selected: false)
-        ]
-        accountDeviceCount = accountDevices.count
-        let session = ChatSession(
-            id: UUID().uuidString,
-            title: "你好",
-            updatedLabel: "刚刚",
-            agentType: "code",
-            workspacePath: "/workspace/BitFun",
-            workspaceName: "BitFun"
-        )
-        remoteSessions = [session]
-        let extraSessions = (1...5).map { index in
-            ChatSession(
-                id: "preview-session-\(index)", title: "Review session \(index)", updatedLabel: "2026-01-01T00:00:00Z",
-                status: index == 1 ? "running" : "idle", agentType: "code",
-                workspacePath: "/workspace/BitFun", workspaceName: "BitFun", deviceKey: "preview-desktop"
-            )
-        }
-        remoteSessions.append(contentsOf: extraSessions)
-        let cachedSession = ChatSession(
-            id: "preview-offline-session", title: "Cached offline session", updatedLabel: "2026-01-01T00:00:00Z",
-            status: "idle", agentType: "code", workspacePath: "/office/project", workspaceName: "Office project", deviceKey: "preview-offline"
-        )
-        let failedSession = ChatSession(
-            id: "preview-failed-session", title: "Cached failed session", updatedLabel: "2026-01-01T00:00:00Z",
-            status: "idle", agentType: "code", workspacePath: "/staging/project", workspaceName: "Staging", deviceKey: "preview-mac"
-        )
-        remoteSessions.append(contentsOf: [cachedSession, failedSession])
-        let previewWorkspace = MobileWorkspaceGroup(path: "/workspace/BitFun", name: "BitFun", selected: true, sessions: remoteSessions.filter { $0.deviceKey == "preview-desktop" }, deviceKey: "preview-desktop")
-        let offlineWorkspace = MobileWorkspaceGroup(path: "/office/project", name: "Office project", selected: false, sessions: [cachedSession], deviceKey: "preview-offline")
-        let failedWorkspace = MobileWorkspaceGroup(path: "/staging/project", name: "Staging", selected: false, sessions: [failedSession], deviceKey: "preview-mac")
-        deviceDirectory = [
-            MobileDeviceDirectoryEntry(id: "preview-desktop", name: "DESKTOP-KM3L4UI", online: true, expanded: true, status: "READY", error: nil, workspaces: [previewWorkspace], sessions: previewWorkspace.sessions),
-            MobileDeviceDirectoryEntry(id: "preview-mac", name: "Studio Mac", online: true, expanded: true, status: "FAILED", error: "REMOTE_UNAVAILABLE", workspaces: [failedWorkspace], sessions: [failedSession]),
-            MobileDeviceDirectoryEntry(id: "preview-offline", name: "Office PC", online: false, expanded: false, status: "READY", error: nil, workspaces: [offlineWorkspace], sessions: [cachedSession])
-        ]
-        workspaceCatalog = [(path: "/workspace/BitFun", name: "BitFun", selected: true)]
-        remoteAssistants = [
-            MobileAssistantOption(path: "/workspace/BitFun/.bitfun/assistants/review", name: "代码审查助手")
-        ]
-        remoteHasMore = true
-        rebuildRemoteWorkspaceGroups()
-        selectedSessionID = session.id
-        messages = [
-            ChatMessage(id: UUID(), role: .user, text: "你好"),
-            ChatMessage(id: UUID(), role: .assistant, text: "这是 BitFun 的远程会话预览。"),
-        ]
-        timelineRows = messages.map(Self.simpleTimelineRow)
-    }
-
-    private func configureTimelinePreview() {
-        configureConnectedPreview()
-        let userID = UUID().uuidString
-        let assistantID = UUID().uuidString
-        let readOne = MobileTimelineTool(
-            id: "preview-read-1", name: "Read", phase: "COMPLETED", kind: "DOCUMENT",
-            operation: "READ_FILE", target: "main.rs", filePath: "computer://src/main.rs",
-            fileLabel: "main.rs", input: "src/main.rs", output: "读取完成", question: nil, questions: [], actions: []
-        )
-        let readTwo = MobileTimelineTool(
-            id: "preview-read-2", name: "Search", phase: "COMPLETED", kind: "SEARCH",
-            operation: "SEARCH_CODE", target: "MobileShellView", filePath: "", fileLabel: "",
-            input: "MobileShellView", output: "找到 4 处结果", question: nil, questions: [], actions: []
-        )
-        let approval = MobileTimelineTool(
-            id: "preview-approval", name: "Bash", phase: "PENDING_CONFIRMATION", kind: "COMMAND",
-            operation: "RUN_COMMAND", target: "pnpm test", filePath: "", fileLabel: "",
-            input: "pnpm test", output: "", question: nil, questions: [], actions: ["APPROVE", "REJECT"]
-        )
-        let question = MobileTimelineTool(
-            id: "preview-question", name: "AskUserQuestion", phase: "PENDING_CONFIRMATION", kind: "QUESTION",
-            operation: "ASK_CONFIRMATION", target: "", filePath: "", fileLabel: "", input: "", output: "",
-            question: "要同时运行远程场景回归吗？", questions: [], actions: ["ANSWER"]
-        )
-        timelineRows = [
-            MobileConversationRow(
-                id: userID, kind: "USER", text: "请检查移动端的消息、工具和文件交互。", thinking: nil,
-                images: [], tools: [], blocks: [], streaming: false, typing: false, pending: false, showRetry: false
-            ),
-            MobileConversationRow(
-                id: assistantID, kind: "ASSISTANT", text: "", thinking: nil, images: [], tools: [],
-                blocks: [
-                    .thinking(id: "preview-thinking", text: "先对照 HarmonyOS 的消息顺序与工具状态，再核对 Android 的交互策略。", streaming: false),
-                    .text(
-                        id: "preview-text",
-                        text: "## 检查结果\n\n消息按共享投影顺序显示，文件可直接打开：[main.rs](computer://src/main.rs)。\n\n- Markdown 与代码块\n- 思考过程与子任务\n- 工具确认、提问和取消\n\n```swift\nlet parity = true\n```",
-                        streaming: false
-                    ),
-                    .tools(id: "preview-tools", tools: [readOne, readTwo, approval, question]),
-                ],
-                streaming: false, typing: false, pending: false, showRetry: false
-            ),
-        ]
-        messages = [
-            ChatMessage(id: UUID(), role: .user, text: "请检查移动端的消息、工具和文件交互。"),
-            ChatMessage(id: UUID(), role: .assistant, text: "检查结果"),
-        ]
     }
 
     func submitPairing(url: String) {
@@ -983,11 +370,6 @@ final class MobileAppModel: ObservableObject {
         connectionPhase = .reconnecting
     }
 
-
-
-
-
-
     func stopSending() {
         if surface == .remote {
             guard remoteSessionSelected else { return }
@@ -996,11 +378,6 @@ final class MobileAppModel: ObservableObject {
             coreAdapter?.cancelGeneralChat()
         }
     }
-
-
-
-
-
 
     func retryMessage(_ text: String) {
         let normalized = text.trimmingCharacters(in: .whitespacesAndNewlines)
@@ -1015,11 +392,6 @@ final class MobileAppModel: ObservableObject {
             send()
         }
     }
-
-
-
-
-
 
     func renameSelectedSession(_ title: String) {
         let normalized = title.trimmingCharacters(in: .whitespacesAndNewlines)
@@ -1050,7 +422,6 @@ final class MobileAppModel: ObservableObject {
         localSessionSelected = false
     }
 
-
     func showUploadedFiles() {
         let count = composerImages.count
         showToast(
@@ -1069,7 +440,6 @@ final class MobileAppModel: ObservableObject {
         }
     }
 
-
     private func apply(pairingState state: PairingUiState, generation: UInt64) {
         guard !localActionPreview, generation == pairingGeneration,
               remoteExpectedDeviceKey == nil || remoteExpectedDeviceKey == "pairing" || pairingIntentInFlight else { return }
@@ -1077,7 +447,7 @@ final class MobileAppModel: ObservableObject {
         if let failed = state as? PairingUiStateFailed {
             pairingBusy = false
             pairingIntentInFlight = false
-            pairingError = pairingErrorMessage(failed.failure)
+            pairingError = PairingFailureCopy.message(failed.failure, localized: localized)
             let healthyConnected: Bool
             switch connectionPhase {
             case .connected: healthyConnected = remoteConnected
@@ -1150,53 +520,5 @@ final class MobileAppModel: ObservableObject {
             }
             pairingSheetOpen = false
         }
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-    private func pairingErrorMessage(_ failure: PairingFailure) -> String {
-        if let remote = failure.remoteMessage?.trimmingCharacters(in: .whitespacesAndNewlines), !remote.isEmpty {
-            return remote
-        }
-        switch failure.reason.name {
-        case "PAIRING_LINK_EMPTY", "PAIRING_LINK_INCOMPLETE", "PAIRING_LINK_UNDECODABLE", "PAIRING_LINK_KEY_UNUSABLE":
-            return localized("连接链接无效，请重新扫描或粘贴桌面端链接")
-        case "ACCOUNT_USERNAME_REQUIRED":
-            return localized("请输入桌面端账号")
-        case "ACCOUNT_PASSWORD_REQUIRED":
-            return localized("请输入桌面端密码")
-        case "REJECTED", "DESKTOP_REJECTED":
-            return localized("桌面端拒绝了这次连接")
-        case "ROOM_NOT_FOUND":
-            return localized("找不到桌面端房间，请确认桌面端仍在等待连接")
-        case "RATE_LIMITED", "TOO_MANY_ATTEMPTS":
-            return localized("尝试次数过多，请稍后再试")
-        case "RELAY_UNAVAILABLE", "NETWORK_UNREACHABLE":
-            return localized("网络不可用，请检查手机与桌面端的网络")
-        case "TIMEOUT":
-            return localized("连接超时，请重新尝试")
-        case "PROTOCOL_MISMATCH":
-            return localized("桌面端版本不兼容，请升级后重试")
-        default:
-            return localized("连接失败，请检查桌面端链接")
-        }
-    }
-
-}
-
-
-private extension Array where Element == String {
-    func value(after flag: String) -> String? {
-        guard let position = firstIndex(of: flag), position < self.index(before: endIndex) else { return nil }
-        return self[self.index(after: position)]
     }
 }
